Added code to replicate SSL problem.
