+++ 1_ClientCredentials/Quickstart1_ClientCredentials.sln
@@ -12,10 +12,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "QuickstartIdentityServer", "src\QuickstartIdentityServer\QuickstartIdentityServer.xproj", "{74894537-E514-4087-B6E1-B9D77A2ED34C}"
 EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Api", "src\Api\Api.xproj", "{7059F655-BBD7-4CA3-9A76-2FF198996DA8}"
-EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Client", "src\Client\Client.xproj", "{B707F1AF-DD46-45D6-B4C5-BC2E9A3FB9B7}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -26,21 +22,11 @@ Global
 		{74894537-E514-4087-B6E1-B9D77A2ED34C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{74894537-E514-4087-B6E1-B9D77A2ED34C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{74894537-E514-4087-B6E1-B9D77A2ED34C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7059F655-BBD7-4CA3-9A76-2FF198996DA8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7059F655-BBD7-4CA3-9A76-2FF198996DA8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7059F655-BBD7-4CA3-9A76-2FF198996DA8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7059F655-BBD7-4CA3-9A76-2FF198996DA8}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B707F1AF-DD46-45D6-B4C5-BC2E9A3FB9B7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B707F1AF-DD46-45D6-B4C5-BC2E9A3FB9B7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B707F1AF-DD46-45D6-B4C5-BC2E9A3FB9B7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B707F1AF-DD46-45D6-B4C5-BC2E9A3FB9B7}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{74894537-E514-4087-B6E1-B9D77A2ED34C} = {2408EC7B-2E06-46A5-B6AC-D4E985E30CBB}
-		{7059F655-BBD7-4CA3-9A76-2FF198996DA8} = {2408EC7B-2E06-46A5-B6AC-D4E985E30CBB}
-		{B707F1AF-DD46-45D6-B4C5-BC2E9A3FB9B7} = {2408EC7B-2E06-46A5-B6AC-D4E985E30CBB}
 	EndGlobalSection
 EndGlobal
