--- 1_ClientCredentials/src/Api/project.json
@@ -1,58 +1,0 @@
-﻿{
-  "dependencies": {
-    "Microsoft.NETCore.App": {
-      "version": "1.0.1",
-      "type": "platform"
-    },
-    "Microsoft.AspNetCore.Mvc.Core": "1.0.*",
-    "Microsoft.AspNetCore.Mvc.Formatters.Json": "1.0.*",
-    "Microsoft.AspNetCore.Server.IISIntegration": "1.0.*",
-    "Microsoft.AspNetCore.Server.Kestrel": "1.0.*",
-    "Microsoft.Extensions.Configuration.EnvironmentVariables": "1.0.*",
-    "Microsoft.Extensions.Configuration.FileExtensions": "1.0.*",
-    "Microsoft.Extensions.Configuration.Json": "1.0.*",
-    "Microsoft.Extensions.Logging": "1.0.*",
-    "Microsoft.Extensions.Logging.Console": "1.0.*",
-    "Microsoft.Extensions.Logging.Debug": "1.0.*",
-    "Microsoft.Extensions.Options.ConfigurationExtensions": "1.0.*",
-
-    "IdentityServer4.AccessTokenValidation": "1.0.1"
-  },
-
-  "tools": {
-    "Microsoft.AspNetCore.Server.IISIntegration.Tools": "1.0.0-preview2-final"
-  },
-
-  "frameworks": {
-    "netcoreapp1.0": {
-      "imports": [
-        "dotnet5.6",
-        "portable-net45+win8"
-      ]
-    }
-  },
-
-  "buildOptions": {
-    "emitEntryPoint": true,
-    "preserveCompilationContext": true
-  },
-
-  "runtimeOptions": {
-    "configProperties": {
-      "System.GC.Server": true
-    }
-  },
-
-  "publishOptions": {
-    "include": [
-      "wwwroot",
-      "**/*.cshtml",
-      "appsettings.json",
-      "web.config"
-    ]
-  },
-
-  "scripts": {
-    "postpublish": [ "dotnet publish-iis --publish-folder %publish:OutputPath% --framework %publish:FullTargetFramework%" ]
-  }
-}
