+++ 1_ClientCredentials/src/QuickstartIdentityServer/Startup.cs
@@ -2,6 +2,8 @@
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
 using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 
@@ -9,14 +11,31 @@ namespace QuickstartIdentityServer
 {
     public class Startup
     {
-        public void ConfigureServices(IServiceCollection services)
+		public Startup(IHostingEnvironment env)
+		{
+			var builder = new ConfigurationBuilder()
+				.SetBasePath(env.ContentRootPath)
+				.AddJsonFile("appsettings.json", optional: true, reloadOnChange: true)
+				.AddJsonFile($"appsettings.{env.EnvironmentName}.json", optional: true);
+
+			builder.AddEnvironmentVariables();
+			Configuration = builder.Build();
+		}
+
+		public IConfigurationRoot Configuration { get; }
+
+		public void ConfigureServices(IServiceCollection services)
         {
             // configure identity server with in-memory stores, keys, clients and scopes
             services.AddIdentityServer()
                 .AddTemporarySigningCredential()
                 .AddInMemoryApiResources(Config.GetApiResources())
                 .AddInMemoryClients(Config.GetClients());
-        }
+			
+			services.AddMvcCore()
+				.AddAuthorization()
+				.AddJsonFormatters();
+		}
 
         public void Configure(IApplicationBuilder app, ILoggerFactory loggerFactory)
         {
@@ -24,6 +43,19 @@ namespace QuickstartIdentityServer
             app.UseDeveloperExceptionPage();
 
             app.UseIdentityServer();
-        }
+
+			loggerFactory.AddConsole(Configuration.GetSection("Logging"));
+			loggerFactory.AddDebug();
+
+			app.UseIdentityServerAuthentication(new IdentityServerAuthenticationOptions
+			{
+				Authority = "https://localhost:5443",
+				RequireHttpsMetadata = true,
+
+				ApiName = "api1"
+			});
+
+			app.UseMvc();
+		}
     }
 }
