+++ 1_ClientCredentials/src/QuickstartIdentityServer/project.json
@@ -1,4 +1,4 @@
-﻿{
+{
   "dependencies": {
     "Microsoft.NETCore.App": {
       "version": "1.1.*",
@@ -8,7 +8,14 @@
     "Microsoft.AspNetCore.Diagnostics": "1.1.*",
     "Microsoft.AspNetCore.Server.IISIntegration": "1.1.*",
     "Microsoft.AspNetCore.Server.Kestrel": "1.1.*",
-    "Microsoft.Extensions.Logging.Console": "1.1.*"
+    "Microsoft.Extensions.Logging.Console": "1.1.*",
+    "Microsoft.AspNetCore.Server.Kestrel.Https": "1.1.0",
+    "Microsoft.Extensions.Configuration.FileExtensions": "1.1.0",
+    "Microsoft.Extensions.Configuration.Json": "1.1.0",
+    "Microsoft.AspNetCore.Mvc.Core": "1.1.1",
+    "Microsoft.AspNetCore.Mvc.Formatters.Json": "1.1.1",
+    "IdentityServer4.AccessTokenValidation": "1.0.5",
+    "Microsoft.Extensions.Logging.Debug": "1.1.0"
   },
 
   "tools": {
